+++ tests/lib/forms-schema.test.js
@@ -1042,6 +1042,70 @@ test('should throw error if minNumber is not provided for number element with is
   expect(error.message).toContain('"Form Element - Minimum Number" is required')
 })
 
+test('should error if number min is not an integer when number isInteger', () => {
+  const { error } = Joi.validate(
+    {
+      id: 1,
+      name: 'Inspection',
+      formsAppEnvironmentId: 1,
+      formsAppIds: [1],
+      organisationId: '59cc888b8969af000fb50ddb',
+      postSubmissionAction: 'FORMS_LIBRARY',
+      submissionEvents: [],
+      tags: [],
+      elements: [
+        {
+          id: '84375ac0-9a0e-11e8-8fc5-63e99eca0edb',
+          name: 'test',
+          type: 'number',
+          label: 'test',
+          required: false,
+          minNumber: 33.5,
+          conditionallyShow: false,
+          isInteger: true,
+        },
+      ],
+    },
+    formSchema,
+  )
+
+  expect(error.message).toContain(
+    '"Form Element - Minimum Number" must be an integer',
+  )
+})
+
+test('should error if number max is not an integer when number isInteger', () => {
+  const { error } = Joi.validate(
+    {
+      id: 1,
+      name: 'Inspection',
+      formsAppEnvironmentId: 1,
+      formsAppIds: [1],
+      organisationId: '59cc888b8969af000fb50ddb',
+      postSubmissionAction: 'FORMS_LIBRARY',
+      submissionEvents: [],
+      tags: [],
+      elements: [
+        {
+          id: '84375ac0-9a0e-11e8-8fc5-63e99eca0edb',
+          name: 'test',
+          type: 'number',
+          label: 'test',
+          required: false,
+          maxNumber: 33.5,
+          conditionallyShow: false,
+          isInteger: true,
+        },
+      ],
+    },
+    formSchema,
+  )
+
+  expect(error.message).toContain(
+    '"Form Element - Maximum Number" must be an integer',
+  )
+})
+
 // Looks like this is impossible with the current Joi version
 // https://github.com/hapijs/joi/issues/1685
 test.skip('should throw error if maxNumber is not provided for number element with isSlider as true', () => {
@@ -4967,5 +5031,40 @@ describe('submission event conditional logic', () => {
     )
 
     expect(result.error).toBeNull()
+  })
+
+  test('should reject number element with decimal default value when isInteger set to true', () => {
+    const result = Joi.validate(
+      {
+        id: 1,
+        name: 'conditionally show element via number input',
+        description: 'string',
+        formsAppEnvironmentId: 1,
+        formsAppIds: [1],
+        organisationId: 'ORGANISATION_00000000001',
+        postSubmissionAction: 'FORMS_LIBRARY',
+        isMultiPage: false,
+        elements: [
+          {
+            id: 'b941ea2d-965c-4d40-8c1d-e5a231fc18b1',
+            name: 'Numbers',
+            label: 'Numbers',
+            type: 'number',
+            required: false,
+            defaultValue: 3.2,
+            isInteger: true,
+          },
+        ],
+        isAuthenticated: true,
+      },
+      formSchema,
+      {
+        abortEarly: false,
+      },
+    )
+
+    expect(result.error.message).toContain(
+      '"Form Element - Default Value" must be an integer',
+    )
   })
 })