+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - minLength/maxLength props for text & textarea elements
+- `isInteger` to `number` element type
 
 ## 0.7.5 (2020-11-05)